backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
PyJWT==2.8.0
Flasgger==0.9.7.1
Werkzeug==2.3.7
python-dotenv==1.0.1
